Moving tests to test-prefixed namespace
