--- tests/phpunit/Civi/API/V4/Action/UpdateCustomValueTest.php
@@ -1,55 +1,0 @@
-<?php
-namespace Civi\API\V4\Action;
-
-use Civi\API\V4\Entity\Contact;
-use Civi\API\V4\Entity\CustomField;
-use Civi\API\V4\Entity\CustomGroup;
-use \CRM_Core_BAO_CustomValueTable as CustomValueTable;
-
-/**
- * @group headless
- */
-class UpdateCustomValueTest extends BaseCustomValueTest {
-
-  public function testGetWithCustomData() {
-
-    $customGroup = CustomGroup::create()
-      ->setCheckPermissions(FALSE)
-      ->setValue('name', 'MyContactFields')
-      ->setValue('extends', 'Contact')
-      ->execute()
-      ->getArrayCopy();
-
-    CustomField::create()
-      ->setCheckPermissions(FALSE)
-      ->setValue('label', 'FavColor')
-      ->setValue('custom_group_id', $customGroup['id'])
-      ->setValue('html_type', 'Text')
-      ->setValue('data_type', 'String')
-      ->execute();
-
-    $contactId = Contact::create()
-      ->setCheckPermissions(FALSE)
-      ->setValue('first_name', 'Red')
-      ->setValue('last_name', 'Tester')
-      ->setValue('contact_type', 'Individual')
-      ->setValue('MyContactFields.FavColor', 'Red')
-      ->execute()
-      ->getArrayCopy()['id'];
-
-    Contact::create()
-      ->setCheckPermissions(FALSE)
-      ->setValue('id', $contactId)
-      ->setValue('first_name', 'Red')
-      ->setValue('last_name', 'Tester')
-      ->setValue('contact_type', 'Individual')
-      ->setValue('MyContactFields.FavColor', 'Blue')
-      ->execute();
-
-    $result = CustomValueTable::getEntityValues($contactId, 'Contact');
-
-    $this->assertEquals(1, count($result));
-    $this->assertContains('Blue', $result);
-  }
-
-}
