+++ tests/phpunit/DataSets/SingleContactWithPhoneEmailAndActivities.json
@@ -1,0 +1,86 @@
+{
+  "Contact": [
+    {
+      "first_name": "Single",
+      "last_name": "Contact",
+      "contact_type": "Individual",
+      "@ref": "test_contact_1"
+    }
+  ],
+  "OptionGroup": [
+    {
+      "name": "activity_type"
+    },
+    {
+      "name": "activity_contacts"
+    }
+  ],
+  "OptionValue": [
+    {
+      "label": "Did Something Amazing",
+      "name": "did_something_great",
+      "option_group": "activity_type"
+    },
+    {
+      "label": "Did That Other Thing",
+      "name": "another_activity_type",
+      "option_group": "activity_type"
+    },
+    {
+      "label": "Activity Source",
+      "name": "Activity Source",
+      "option_group": "activity_contacts"
+    },
+    {
+      "label": "Activity Assignees",
+      "name": "Activity Assignees",
+      "option_group": "activity_contacts"
+    },
+    {
+      "label": "Activity Targets",
+      "name": "Activity Targets",
+      "option_group": "activity_contacts"
+    }
+  ],
+  "Activity": [
+    {
+      "subject": "Won A Nobel Prize",
+      "activity_type": "did_something_great",
+      "source_contact_id": "@ref test_contact_1.id",
+      "@ref": "test_activity_1"
+    },
+    {
+      "subject": "Cleaned The House",
+      "activity_type": "another_activity_type",
+      "source_contact_id": "@ref test_contact_1.id",
+      "assignee_contact_id": [
+        "@ref test_contact_1.id"
+      ],
+      "@ref": "test_activity_2"
+    }
+  ],
+  "Phone": [
+    {
+      "contact_id": "@ref test_contact_1.id",
+      "phone": "+1111111111111",
+      "location_type_id": 1
+    },
+    {
+      "contact_id": "@ref test_contact_1.id",
+      "phone": "+2222222222222",
+      "location_type_id": 2
+    }
+  ],
+  "Email": [
+    {
+      "contact_id": "@ref test_contact_1.id",
+      "email": "test_contact_home@fakedomain.com",
+      "location_type_id": 1
+    },
+    {
+      "contact_id": "@ref test_contact_1.id",
+      "email": "test_contact_work@fakedomain.com",
+      "location_type_id": 2
+    }
+  ]
+}
