+++ tests/phpunit/Query/Api4SelectQueryComplexJoinTest.php
@@ -1,0 +1,72 @@
+<?php
+
+namespace Civi\Test\API\V4\Query;
+
+use Civi\API\V4\Query\Api4SelectQuery;
+use Civi\Test\API\V4\Traits\TestDataLoaderTrait;
+use Civi\Test\API\V4\UnitTestCase;
+
+/**
+ * @group headless
+ */
+class Api4SelectQueryComplexJoinTest extends UnitTestCase {
+
+  use TestDataLoaderTrait;
+
+  public function setUpHeadless() {
+    $relatedTables = array(
+      'civicrm_contact',
+      'civicrm_option_group',
+      'civicrm_option_value',
+      'civicrm_email',
+      'civicrm_phone',
+      'civicrm_activity',
+      'civicrm_activity_contact',
+    );
+    $this->cleanup(array('tablesToTruncate' => $relatedTables));
+    $this->loadDataSet('SingleContactWithPhoneEmailAndActivities');
+    return parent::setUpHeadless();
+  }
+
+  public function testWithComplexRelatedEntitySelect() {
+    $query = new Api4SelectQuery('Contact', FALSE);
+    $query->select[] = 'id';
+    $query->select[] = 'display_name';
+    $query->select[] = 'phones.phone';
+    $query->select[] = 'emails.email';
+    $query->select[] = 'emails.location_type.name';
+    $query->select[] = 'created_activities.contact_id';
+    $query->select[] = 'created_activities.activity.subject';
+    $query->select[] = 'created_activities.activity.activity_type.name';
+    $query->where[] = array('first_name', '=', 'Single');
+    $results = $query->run();
+
+    $testActivities = [
+      $this->getReference('test_activity_1'),
+      $this->getReference('test_activity_2'),
+    ];
+    $activitySubjects = array_column($testActivities, 'subject');
+
+    $this->assertCount(1, $results);
+    $firstResult = array_shift($results);
+    $this->assertArrayHasKey('created_activities', $firstResult);
+    $firstCreatedActivity = array_shift($firstResult['created_activities']);
+    $this->assertArrayHasKey('activity', $firstCreatedActivity);
+    $firstActivity = $firstCreatedActivity['activity'];
+    $this->assertContains($firstActivity['subject'], $activitySubjects);
+    $this->assertArrayHasKey('activity_type', $firstActivity);
+    $activityType = $firstActivity['activity_type'];
+    $this->assertArrayHasKey('name', $activityType);
+  }
+
+  public function testWithSelectOfOrphanDeepValues() {
+    $query = new Api4SelectQuery('Contact', FALSE);
+    $query->select[] = 'id';
+    $query->select[] = 'first_name';
+    $query->select[] = 'emails.location_type.name'; // emails not selected
+    $results = $query->run();
+    $firstResult = array_shift($results);
+
+    $this->assertEmpty($firstResult['emails']);
+  }
+}
