+++ tests/phpunit/Utils/TestV4ReflectionBase.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Civi\Test\API\V4\Utils;
+
+/**
+ * Class TestV4ReflectionBase
+ *
+ * This is the base class.
+ *
+ * @internal
+ */
+class TestV4ReflectionBase {
+  /**
+   * This is the foo property.
+   *
+   * In general, you can do nothing with it.
+   *
+   * @var array
+   */
+  public $foo = array();
+
+}
