+++ tests/phpunit/Utils/TestV4ReflectionGrandchild.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Civi\Test\API\V4\Utils;
+
+use Civi\Test\API\V4\Utils;
+
+
+/**
+ * Grandchild class
+ *
+ * This is an extended description.
+ *
+ * There is a line break in this description.
+ *
+ * @inheritDoc
+ */
+class TestV4ReflectionGrandchild extends Utils\TestV4ReflectionChild {
+
+}
